# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Clinic Booking API",
        "description": "API for managing online clinic bookings, including patients, practitioners, and admins.",
        "contact": {
            "email": "support@onlineclinicbooking.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/practitioners": {
            "get": {
                "tags": [
                    "Practitioners"
                ],
                "summary": "Get List of Practitioners",
                "description": "Retrieve a list of all practitioners",
                "operationId": "92dc8c1b1cce75bbc1ec59c32355cf0f",
                "responses": {
                    "200": {
                        "description": "List of practitioners",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "practitioner_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "full_name": {
                                                "type": "string",
                                                "example": "Dr. Smith"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "dr.smith@example.com"
                                            },
                                            "specialty": {
                                                "type": "string",
                                                "example": "Cardiology"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Practitioners"
                ],
                "summary": "Create a New Practitioner",
                "description": "Create a new practitioner record",
                "operationId": "bc1ed24294a46c0484f8b2cec14aec2b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "full_name"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "practitioner@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "full_name": {
                                        "type": "string",
                                        "example": "Dr. Smith"
                                    },
                                    "specialty": {
                                        "type": "string",
                                        "example": "Cardiology"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Practitioner created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Practitioner created successfully"
                                        },
                                        "practitioner": {
                                            "properties": {
                                                "practitioner_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "full_name": {
                                                    "type": "string",
                                                    "example": "Dr. Smith"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "practitioner@example.com"
                                                },
                                                "specialty": {
                                                    "type": "string",
                                                    "example": "Cardiology"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email has already been taken."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email has already been taken."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter the token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}